northd: Add LR option to always evaluate policies.
Add LR option "fallback_to_routing_policy" that, when set to true,
changes router behavior to always evaluate routing policies, even if
there was no match in the routing table (otherwise it would simply drop
the packet).

If the option is set, we adjust the lflows in such a way that the
decision to drop the packet is delayed until we reach the lr_in_policy
pipeline.

The change takes advantage of the fact that the IPv4 and IPv6 src ip
and nexthop registers share bits.  Thus, at least at the moment, we
don't need to use another register to save the lookup result (match or
no match) from the lr_in_ip_routing pipeline.

Having this option is useful for routers that do not have a default
route and one still wants to make use of the more advanced matching
capabilities of routing policies.

Signed-off-by: Martin Morgenstern <martin.morgenstern@cloudandheat.com>